fix(coding-agent): break the mission renderer import cycle
Registering the mission tool made tools/renderers.ts import
missions/tools/mission-tool.ts, which reaches MissionRuntime and from
there the task executor, closing a loop back into tools. The renderer
table then evaluated before mission-tool.ts finished, so
missionToolRenderer was read in its temporal dead zone and every suite
importing the registry died with a ReferenceError.

Split the presentation half into missions/tools/mission-render.ts,
which imports only TUI and type leaves. mission-tool.ts takes
MissionToolDetails from it and the renderer table depends on the leaf.
Also update the HIDDEN_TOOLS assertion, which pinned the exact key set.
