+++ packages/coding-agent/src/missions/index.ts
@@ -1,5 +1,6 @@
 export * from "./runtime";
 export * from "./state";
+export * from "./tools/mission-render";
 export * from "./tools/mission-tool";
 export * from "./types";
 export * from "./workspace";
