+++ packages/coding-agent/src/missions/tools/mission-tool.ts
@@ -11,21 +11,14 @@
  * `/mission` or RPC, never model tool ops, so they are intentionally absent.
  */
 import type { AgentTool, AgentToolContext, AgentToolResult, AgentToolUpdateCallback } from "@oh-my-pi/pi-agent-core";
-import type { Component } from "@oh-my-pi/pi-tui";
-import { Text } from "@oh-my-pi/pi-tui";
-import { isRecord, prompt } from "@oh-my-pi/pi-utils";
+import { prompt } from "@oh-my-pi/pi-utils";
 import { type } from "arktype";
-import type { RenderResultOptions } from "../../extensibility/custom-tools/types";
-import type { Theme } from "../../modes/theme/theme";
 import missionDescription from "../../prompts/tools/mission.md" with { type: "text" };
 import type { ToolSession } from "../../tools";
-import type { OutputMeta } from "../../tools/output-meta";
-import { formatErrorDetail, replaceTabs, TRUNCATE_LENGTHS } from "../../tools/render-utils";
 import { ToolError } from "../../tools/tool-errors";
 import { toolResult } from "../../tools/tool-result";
-import { framedBlock, renderStatusLine, truncateToWidth } from "../../tui";
 import { MissionRuntimeError } from "../runtime";
-import type { MissionHandoffDecision, MissionStatus } from "../types";
+import type { MissionToolDetails } from "./mission-render";
 
 const remediationFeatureSchema = type({
 	id: type("string").describe("stable remediation feature id"),
@@ -92,13 +85,6 @@ const missionSchema = type({ op: "'get'" })
 	.or({ op: "'revise_pending'", add_features: remediationFeatureSchema.array() });
 
 export type MissionToolInput = typeof missionSchema.infer;
-
-export type MissionToolDetails =
-	| { op: "get"; status: MissionStatus | "none"; meta?: OutputMeta }
-	| { op: "set_plan"; status: MissionStatus; milestoneCount: number; featureCount: number; meta?: OutputMeta }
-	| { op: "run_next"; handoff: "implementation" | "validation" | "none"; meta?: OutputMeta }
-	| { op: "resolve_handoff"; decision: MissionHandoffDecision; status: MissionStatus; meta?: OutputMeta }
-	| { op: "revise_pending"; status: MissionStatus; addedFeatures: number; meta?: OutputMeta };
 
 export class MissionTool implements AgentTool<typeof missionSchema, MissionToolDetails> {
 	readonly name = "mission";
@@ -208,117 +194,3 @@ export class MissionTool implements AgentTool<typeof missionSchema, MissionToolD
 		}
 	}
 }
-
-interface MissionRenderArgs {
-	op?: MissionToolInput["op"];
-	decision?: string;
-	message_to_worker?: string;
-}
-
-/**
- * Streamed/partial tool calls hand the renderer arbitrary decoded JSON, so every
- * field is treated as unknown and narrowed before use. Model-supplied strings are
- * tab-replaced and width-capped so they cannot punch layout holes in the TUI.
- */
-function coerceMissionRenderArgs(args: unknown): MissionRenderArgs {
-	if (!isRecord(args)) return {};
-	const op = args.op;
-	return {
-		op:
-			op === "get" || op === "set_plan" || op === "run_next" || op === "resolve_handoff" || op === "revise_pending"
-				? op
-				: undefined,
-		decision: typeof args.decision === "string" ? args.decision : undefined,
-		message_to_worker: typeof args.message_to_worker === "string" ? args.message_to_worker : undefined,
-	};
-}
-
-function sanitizeMeta(value: string): string {
-	return truncateToWidth(replaceTabs(value), TRUNCATE_LENGTHS.TITLE);
-}
-
-function describeOp(op: MissionRenderArgs["op"]): string {
-	switch (op) {
-		case "get":
-			return "get";
-		case "set_plan":
-			return "set plan";
-		case "run_next":
-			return "run next";
-		case "resolve_handoff":
-			return "resolve handoff";
-		case "revise_pending":
-			return "revise pending";
-		default:
-			return op ?? "?";
-	}
-}
-
-export const missionToolRenderer = {
-	renderCall(rawArgs: unknown, _options: RenderResultOptions, uiTheme: Theme): Component {
-		const args = coerceMissionRenderArgs(rawArgs);
-		const description = describeOp(args.op);
-		const meta: string[] = [];
-		if (args.op === "resolve_handoff" && args.decision) {
-			meta.push(uiTheme.fg("muted", sanitizeMeta(args.decision)));
-		}
-		const message = args.message_to_worker?.trim();
-		if (message) {
-			meta.push(uiTheme.italic(uiTheme.fg("muted", `"${sanitizeMeta(message)}"`)));
-		}
-		return new Text(renderStatusLine({ icon: "pending", title: "Mission", description, meta }, uiTheme), 0, 0);
-	},
-
-	renderResult(
-		result: { content: Array<{ type: string; text?: string }>; details?: MissionToolDetails; isError?: boolean },
-		_options: RenderResultOptions,
-		uiTheme: Theme,
-		rawArgs?: unknown,
-	): Component {
-		const fallbackText = result.content?.find(part => part.type === "text")?.text ?? "";
-		const details = result.details;
-		const description = describeOp(details?.op ?? coerceMissionRenderArgs(rawArgs).op);
-
-		if (result.isError) {
-			const header = renderStatusLine({ icon: "error", title: "Mission", description }, uiTheme);
-			return framedBlock(uiTheme, width => ({
-				header,
-				sections: [{ lines: formatErrorDetail(fallbackText || "Mission tool failed", uiTheme).split("\n") }],
-				state: "error",
-				borderColor: "error",
-				width,
-			}));
-		}
-
-		const meta: string[] = [];
-		if (details) {
-			switch (details.op) {
-				case "get":
-					meta.push(uiTheme.fg("muted", details.status));
-					break;
-				case "set_plan":
-					meta.push(uiTheme.fg("muted", details.status));
-					meta.push(uiTheme.fg("muted", `${details.milestoneCount}m/${details.featureCount}f`));
-					break;
-				case "run_next":
-					meta.push(uiTheme.fg("muted", details.handoff));
-					break;
-				case "resolve_handoff":
-					meta.push(uiTheme.fg("muted", details.decision));
-					meta.push(uiTheme.fg("muted", details.status));
-					break;
-				case "revise_pending":
-					meta.push(uiTheme.fg("muted", details.status));
-					meta.push(uiTheme.fg("muted", `+${details.addedFeatures}`));
-					break;
-				default: {
-					const _exhaustive: never = details;
-					void _exhaustive;
-					break;
-				}
-			}
-		}
-
-		return new Text(renderStatusLine({ icon: "success", title: "Mission", description, meta }, uiTheme), 0, 0);
-	},
-};
